pyproject.toml
[tool.poetry]
name = "chat-inteligente"
version = "0.1.0"
description = ""
authors = ["victor gil <arklon1975@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
